test_fullscan() 通った

+++ apllodb-server/tests/query.rs
@@ -27,7 +27,7 @@ async fn test_fullscan() {
             "SELECT id, age FROM people",
             StepRes::OkQuery(Box::new(|records| {
                 let mut records = records
-                    .sorted_by_key(|r| r.get::<i32>(&field("people", "id")).unwrap().unwrap());
+                    .sorted_by_key(|r| r.get::<i64>(&field("people", "id")).unwrap().unwrap());
                 assert_eq!(records.next(), Some(T_PEOPLE_R1.clone()));
                 assert_eq!(records.next(), Some(T_PEOPLE_R2.clone()));
                 assert_eq!(records.next(), Some(T_PEOPLE_R3.clone()));
@@ -39,7 +39,7 @@ async fn test_fullscan() {
             "SELECT people_id, height FROM body",
             StepRes::OkQuery(Box::new(|records| {
                 let mut records =
-                    records.sorted_by_key(|r| r.get::<i32>(&field("body", "id")).unwrap().unwrap());
+                    records.sorted_by_key(|r| r.get::<i64>(&field("body", "id")).unwrap().unwrap());
                 assert_eq!(records.next(), Some(T_BODY_R1.clone()));
                 assert_eq!(records.next(), Some(T_BODY_R3.clone()));
                 assert!(records.next().is_none());
@@ -50,7 +50,7 @@ async fn test_fullscan() {
             "SELECT people_id, kind, age FROM pet",
             StepRes::OkQuery(Box::new(|records| {
                 let mut records =
-                    records.sorted_by_key(|r| r.get::<i32>(&field("pet", "id")).unwrap().unwrap());
+                    records.sorted_by_key(|r| r.get::<i64>(&field("pet", "id")).unwrap().unwrap());
                 assert_eq!(records.next(), Some(T_PET_R1.clone()));
                 assert_eq!(records.next(), Some(T_PET_R3_1.clone()));
                 assert_eq!(records.next(), Some(T_PET_R3_2.clone()));
